Move createSvgIcon fuction to utils
