+++ react/components/stormPoint.js
@@ -7,6 +7,7 @@ import {empty_point_obj} from "@/components/storm_point_details"
 import { Tooltip } from 'react-leaflet';
 import { useMediaQuery, useTheme } from '@mui/material';
 import StormPointDetailsSmallScreen from "./storm_point_details_small_screens";
+import { createSvgIconWithText } from "./utils/storm_display_utils";
 
 
 
@@ -100,27 +101,6 @@ export default function StormMarker1({ storm_point_data, setHoverMarker, setIsSt
     );
 }
 
-function createSvgIconWithText(text = "A", fill = "#0077be", size = 40, textColor = 'white') {
-    const fontSize = (size / 2) - 2;
-     
-    const svg = `
-      <svg width="${size}" height="${size}" viewBox="0 0 ${size} ${size}" xmlns="http://www.w3.org/2000/svg">
-        <circle cx="${size / 2}" cy="${size / 2}" r="${size / 2 - 2}" fill="${fill}" 
-                stroke="black" stroke-width="1.5" />
-        <text x="50%" y="50%" dominant-baseline="middle" text-anchor="middle"
-              fill="${textColor}" font-size="${fontSize}" font-family="Arial, sans-serif"
-              >
-          ${text}
-        </text>
-      </svg>
-    `;
-    
-    return L.divIcon({
-      html: svg,
-      className: '',
-      iconSize: [size, size],
-      iconAnchor: [size / 2, size / 2],
-    });
-  };
+
